+++ libs/tendermint/abci/client/client.go
@@ -29,7 +29,6 @@ type Client interface {
 	InfoAsync(types.RequestInfo) *ReqRes
 	SetOptionAsync(types.RequestSetOption) *ReqRes
 	DeliverTxAsync(types.RequestDeliverTx) *ReqRes
-	DeliverTxConcurrently([][]byte, types.DeliverTxContext) []*ReqRes
 	CheckTxAsync(types.RequestCheckTx) *ReqRes
 	QueryAsync(types.RequestQuery) *ReqRes
 	CommitAsync(types.RequestCommit) *ReqRes
