+++ libs/tendermint/abci/types/application.go
@@ -22,7 +22,6 @@ type Application interface {
 	InitChain(RequestInitChain) ResponseInitChain    // Initialize blockchain w validators/other info from TendermintCore
 	BeginBlock(RequestBeginBlock) ResponseBeginBlock // Signals the beginning of a block
 	DeliverTx(RequestDeliverTx) ResponseDeliverTx    // Deliver a tx for full processing
-	DeliverTxConcurrently([][]byte, DeliverTxContext) []*ResponseDeliverTx    // Deliver a tx for full processing
 	EndBlock(RequestEndBlock) ResponseEndBlock       // Signals the end of a block, returns changes to the validator set
 
 	Commit(RequestCommit) ResponseCommit             // Commit the state and return the application Merkle root hash
@@ -59,16 +58,6 @@ func (BaseApplication) DeliverTx(req RequestDeliverTx) ResponseDeliverTx {
 	return ResponseDeliverTx{Code: CodeTypeOK}
 }
 
-func (BaseApplication) 	DeliverTxConcurrently(txs [][]byte, _ DeliverTxContext) []*ResponseDeliverTx    {
-	fmt.Printf(" (BaseApplication) DeliverTxConcurrently\n")
-
-	var res []*ResponseDeliverTx
-	for i := 0; i < len(txs); i++ {
-		res = append(res, &ResponseDeliverTx{Code: CodeTypeOK})
-	}
-
-	return res
-}
 
 func (BaseApplication) CheckTx(req RequestCheckTx) ResponseCheckTx {
 	return ResponseCheckTx{Code: CodeTypeOK}
